+++ flash-core/src/main/java/cn/enilu/flash/service/sales/Sla01Service.java
@@ -18,7 +18,7 @@ public class Sla01Service extends BaseService<Sla01,Long,Sla01Repository>{
 	@Autowired
 	private Sla01Repository sla01Repository;
 	
-	public List<Sla01> findBySla01002(Integer sla01002){
+	public List<Sla01> findBySla01002(Long sla01002){
 		return sla01Repository.findBySla01002(sla01002);
 	}
